move updateDistributionsMetadata into datasetVersion.js
