--- src/app/actions/updateDistributionsMetadata.js
@@ -1,52 +1,0 @@
-import { httpPut } from "@/utils/request/request";
-import { logError } from "@/utils/log/log";
-import { getDistributionPath } from "@/utils/url/url";
-
-/**
- * Updates file metadata for each distribution in a dataset version.
- *
- * @param {object} reqCfg - Request config object generated by SSRequestConfig function.
- * @param {Array<{download_url: string}>} distributions - Array of distribution objects containing download URLs.
- * @param {string} datasetID - The dataset ID.
- * @param {string} editionID - The edition ID.
- * @param {number} versionID - The version number.
- * @returns {Promise<{success: boolean, failures: Array<{download_url: string, status: number|null, error: string|null}>}>}
- *   Resolves with a summary of any failed updates.
- */
-
-export const updateDistributionsMetadata = async (reqCfg, distributions = [], datasetID, editionID, versionID) => {
-    const fileMetadataUpdateRequests = distributions.map(async (distribution) => {
-        const filePath = getDistributionPath(distribution?.download_url);
-
-        if (!filePath) {
-            logError("invalid file path", distribution, null, null);
-            return { success: false, distribution, status: null, error: "invalid file path" };
-        }
-
-        try {
-            const response = await httpPut(reqCfg, `/files/${filePath}`, {
-                content_item: {
-                    dataset_id: datasetID,
-                    edition: editionID,
-                    version: String(versionID),
-                },
-            });
-
-            if (response.status >= 400) {
-                logError("failed to update file metadata", distribution, response.status, null);
-                return { success: false, distribution, status: response.status, error: response.errorMessage };
-            }
-
-            return { success: true, distribution };
-
-        } catch (err) {
-            logError("error updating file metadata", distribution, null, err);
-            return { success: false, distribution, status: null, error: err.toString() };
-        }
-    });
-
-    const fileMetadataUpdateResponses = await Promise.all(fileMetadataUpdateRequests);
-    const failures = fileMetadataUpdateResponses.filter(r => !r.success);
-
-    return { success: failures.length === 0, failures };
-};
